修改行编辑时root目录添加时顺序号错误问题

+++ server/src/main/java/muyu/system/common/beans/BaseResultBean.java
@@ -30,7 +30,7 @@ public class BaseResultBean implements Serializable {
 
     private int code = SUCCESS;
 
-    private long useTime=0;
+    private double useTime=0;
 
 
     public BaseResultBean() {
